standardizing error handling
