+++ internal/server/dto.go
@@ -1,0 +1,11 @@
+package server
+
+
+type ErrorRes struct {
+	Error string `json:"error"`
+}
+
+type OkRes struct {
+	Ok bool `json:"ok"`
+}
+
