+++ internal/server/error.go
@@ -1,0 +1,10 @@
+package server
+
+import (
+	"github.com/labstack/echo"
+)
+
+//todo common error htt phandler to inject into echo
+func errorHttpHandler(err error, c echo.Context) {
+
+}
