+++ internal/server/handlers.go
@@ -14,76 +14,32 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-type ErrorRes struct {
-	Error string `json:"error"`
+
+// Header returns the header of the wrapped response writer
+func (frw *FallbackResponseWriter) Header() http.Header {
+	return frw.WrappedResponseWriter.Header()
 }
 
-type OkRes struct {
-	Ok bool `json:"ok"`
-}
-
-func listUsers(us resources.UserService) echo.HandlerFunc {
-	return func(c echo.Context) error {
-		users := us.GetAllUsers(c.Request().Context())
-
-		return c.JSON(http.StatusOK, users)
+// Write sends bytes to wrapped response writer, in case of FileNotFound
+// It surpresses further writes (concealing the fact though)
+func (frw *FallbackResponseWriter) Write(b []byte) (int, error) {
+	if frw.FileNotFound {
+		return len(b), nil
 	}
+	return frw.WrappedResponseWriter.Write(b)
 }
 
-func createUser(us resources.UserService) echo.HandlerFunc {
-	return func(c echo.Context) error {
-		type request struct {
-			Name string `json:"name" validate:"required"`
-		}
-		type response = resources.User
-		r := new(request)
-		if err := c.Bind(r); err != nil {
-			panic(err)
-		}
-		if err := c.Validate(r); err != nil {
-			return c.JSON(http.StatusBadRequest, ErrorRes{err.Error()})
-		}
-		if !isValidUsername(r.Name) {
-			return c.JSON(http.StatusBadRequest, ErrorRes{invalidUsernameError})
-		}
+// WriteHeader sends statusCode to wrapped response writer
+func (frw *FallbackResponseWriter) WriteHeader(statusCode int) {
 
-		u, err := us.CreateUser(c.Request().Context(), r.Name)
-
-		if err != nil {
-
-			return c.JSON(http.StatusBadRequest, ErrorRes{
-				Error: err.Error(),
-			})
-		}
-
-		return c.JSON(http.StatusOK, response{Name: u.Name})
+	if statusCode == http.StatusNotFound {
+		frw.FileNotFound = true
+		return
 	}
+
+	frw.WrappedResponseWriter.WriteHeader(statusCode)
 }
 
-func deleteUser(us resources.UserService) echo.HandlerFunc {
-	return func(c echo.Context) error {
-
-		type Request struct {
-			Username string `json:"username" validate:"required"`
-		}
-
-		r := new(Request)
-		if err := c.Bind(r); err != nil {
-			return err
-		}
-		if err := c.Validate(r); err != nil {
-			return c.JSON(http.StatusBadRequest, ErrorRes{err.Error()})
-		}
-
-		err := us.DeleteUser(c.Request().Context(), r.Username)
-
-		if err != nil {
-			return c.JSON(http.StatusBadRequest, ErrorRes{Error: err.Error()})
-		}
-
-		return c.JSON(http.StatusOK, OkRes{Ok: true})
-	}
-}
 
 func ListNamespaces(rs resources.ResourcesService) echo.HandlerFunc {
 	return func(c echo.Context) error {
@@ -91,14 +47,19 @@ func ListNamespaces(rs resources.ResourcesService) echo.HandlerFunc {
 			Namespaces []string `json:"namespaces"`
 		}
 
-		names, _ := rs.GetAllNamespaces(c.Request().Context())
+		names, err := rs.GetAllNamespaces(c.Request().Context())
+
+		if err != nil {
+			return err
+		}
+
 		return c.JSON(http.StatusOK, Response{
 			Namespaces: names,
 		})
 	}
 }
 
-func ListRbac(c echo.Context) error {
+func listRbac(c echo.Context) error {
 	ac := c.(*AppContext)
 	type Response struct {
 		ClusterRoles        []rbacv1.ClusterRole        `json:"clusterRoles"`
@@ -109,22 +70,22 @@ func ListRbac(c echo.Context) error {
 
 	clusterRoles, err := ac.Kubeclient.RbacV1().ClusterRoles().List(c.Request().Context(), metav1.ListOptions{})
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, ErrorRes{Error: err.Error()})
+		return err
 	}
 
 	clusterRoleBindings, err := ac.Kubeclient.RbacV1().ClusterRoleBindings().List(c.Request().Context(), metav1.ListOptions{})
 	if err != nil {
-		panic(err.Error())
+		return err
 	}
 
 	roles, err := ac.Kubeclient.RbacV1().Roles("").List(c.Request().Context(), metav1.ListOptions{})
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, ErrorRes{Error: err.Error()})
+		return err
 	}
 
 	roleBindings, err := ac.Kubeclient.RbacV1().RoleBindings("").List(c.Request().Context(), metav1.ListOptions{})
 	if err != nil {
-		panic(err.Error())
+		return err
 	}
 
 	return c.JSON(http.StatusOK, Response{
@@ -135,196 +96,7 @@ func ListRbac(c echo.Context) error {
 	})
 }
 
-func CreateClusterRole(c echo.Context) error {
-	ac := c.(*AppContext)
-	type Request struct {
-		RoleName string              `json:"roleName" validate:"required"`
-		Rules    []rbacv1.PolicyRule `json:"rules" validate:"required"`
-	}
-	r := new(Request)
-	if err := c.Bind(r); err != nil {
-		return err
-	}
-	if err := c.Validate(r); err != nil {
-		return c.JSON(http.StatusBadRequest, ErrorRes{err.Error()})
-	}
 
-	_, err := ac.Kubeclient.RbacV1().ClusterRoles().Create(c.Request().Context(), &rbacv1.ClusterRole{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: r.RoleName,
-		},
-		Rules: r.Rules,
-	}, metav1.CreateOptions{})
-
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, ErrorRes{Error: err.Error()})
-	}
-
-	return c.JSON(http.StatusOK, OkRes{Ok: true})
-}
-
-func CreateRolebinding(c echo.Context) error {
-	ac := c.(*AppContext)
-	type Request struct {
-		RolebindingName string           `json:"rolebindingName" validate:"required"`
-		Namespace       string           `json:"namespace" validate:"required"`
-		Username        string           `json:"generated_for_user" validate:"required"`
-		Subjects        []rbacv1.Subject `json:"subjects" validate:"required"`
-		RoleKind        string           `json:"roleKind" validate:"required"`
-		RoleName        string           `json:"roleName" validate:"required"`
-	}
-	r := new(Request)
-	if err := c.Bind(r); err != nil {
-		return err
-	}
-	if err := c.Validate(r); err != nil {
-		return c.JSON(http.StatusBadRequest, ErrorRes{err.Error()})
-	}
-
-	_, err := ac.Kubeclient.RbacV1().RoleBindings(r.Namespace).Create(c.Request().Context(), &rbacv1.RoleBinding{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      r.RolebindingName,
-			Namespace: r.Namespace,
-			Labels:    map[string]string{"generated_for_user": r.Username},
-		},
-		RoleRef: rbacv1.RoleRef{
-			Kind:     r.RoleKind,
-			Name:     r.RoleName,
-			APIGroup: "rbac.authorization.k8s.io",
-		},
-		Subjects: r.Subjects,
-	}, metav1.CreateOptions{})
-
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, ErrorRes{Error: err.Error()})
-	}
-
-	return c.JSON(http.StatusOK, OkRes{Ok: true})
-}
-
-func createClusterRolebinding(c echo.Context) error {
-	ac := c.(*AppContext)
-	type Request struct {
-		ClusterRolebindingName string           `json:"clusterRolebindingName"`
-		Username               string           `json:"user"`
-		Subjects               []rbacv1.Subject `json:"subjects"`
-		RoleName               string           `json:"roleName"`
-	}
-	r := new(Request)
-	if err := c.Bind(r); err != nil {
-		return err
-	}
-
-	ac.Kubeclient.RbacV1().ClusterRoleBindings().Create(c.Request().Context(), &rbacv1.ClusterRoleBinding{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:   r.ClusterRolebindingName,
-			Labels: map[string]string{"generated_for_user": r.Username},
-		},
-		RoleRef: rbacv1.RoleRef{
-			Kind:     "ClusterRole",
-			Name:     r.RoleName,
-			APIGroup: "rbac.authorization.k8s.io",
-		},
-		Subjects: r.Subjects,
-	}, metav1.CreateOptions{})
-
-	return c.JSON(http.StatusOK, OkRes{Ok: true})
-}
-
-func deleteClusterRole(c echo.Context) error {
-	ac := c.(*AppContext)
-	type Request struct {
-		RoleName string `json:"roleName" validate:"required"`
-	}
-
-	r := new(Request)
-	if err := c.Bind(r); err != nil {
-		return err
-	}
-	if err := c.Validate(r); err != nil {
-		return c.JSON(http.StatusBadRequest, ErrorRes{err.Error()})
-	}
-
-	err := ac.Kubeclient.RbacV1().ClusterRoles().Delete(c.Request().Context(), r.RoleName, metav1.DeleteOptions{})
-
-	if err != nil {
-		panic(err.Error())
-	}
-
-	return c.JSON(http.StatusOK, OkRes{Ok: true})
-}
-
-func deleteClusterRolebinding(c echo.Context) error {
-	ac := c.(*AppContext)
-	type Request struct {
-		RolebindingName string `json:"rolebindingName"`
-	}
-
-	r := new(Request)
-	if err := c.Bind(r); err != nil {
-		return err
-	}
-	if err := c.Validate(r); err != nil {
-		return c.JSON(http.StatusBadRequest, ErrorRes{err.Error()})
-	}
-
-	err := ac.Kubeclient.RbacV1().ClusterRoleBindings().Delete(c.Request().Context(), r.RolebindingName, metav1.DeleteOptions{})
-
-	if err != nil {
-		panic(err.Error())
-	}
-
-	return c.JSON(http.StatusOK, OkRes{Ok: true})
-}
-
-func deleteRole(c echo.Context) error {
-	ac := c.(*AppContext)
-
-	type Request struct {
-		RoleName  string `json:"roleName" validate:"required"`
-		Namespace string `json:"namespace" validate:"required"`
-	}
-
-	r := new(Request)
-	if err := c.Bind(r); err != nil {
-		return err
-	}
-	if err := c.Validate(r); err != nil {
-		return c.JSON(http.StatusBadRequest, ErrorRes{err.Error()})
-	}
-
-	err := ac.Kubeclient.RbacV1().Roles(r.Namespace).Delete(c.Request().Context(), r.RoleName, metav1.DeleteOptions{})
-
-	if err != nil {
-		panic(err.Error())
-	}
-
-	return c.JSON(http.StatusOK, OkRes{Ok: true})
-}
-
-func deleteRolebinding(c echo.Context) error {
-	ac := c.(*AppContext)
-	type Request struct {
-		RolebindingName string `json:"rolebindingName" validate:"required"`
-		Namespace       string `json:"namespace" validate:"required"`
-	}
-
-	r := new(Request)
-	if err := c.Bind(r); err != nil {
-		return err
-	}
-	if err := c.Validate(r); err != nil {
-		return c.JSON(http.StatusBadRequest, ErrorRes{err.Error()})
-	}
-
-	err := ac.Kubeclient.RbacV1().RoleBindings(r.Namespace).Delete(c.Request().Context(), r.RolebindingName, metav1.DeleteOptions{})
-
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, ErrorRes{err.Error()})
-	}
-
-	return c.JSON(http.StatusOK, OkRes{Ok: true})
-}
 
 func createKubeconfig(clusterName, clusterControlPlaceAddress string) echo.HandlerFunc {
 	type Request struct {
@@ -362,32 +134,8 @@ type (
 	}
 )
 
-// Header returns the header of the wrapped response writer
-func (frw *FallbackResponseWriter) Header() http.Header {
-	return frw.WrappedResponseWriter.Header()
-}
 
-// Write sends bytes to wrapped response writer, in case of FileNotFound
-// It surpresses further writes (concealing the fact though)
-func (frw *FallbackResponseWriter) Write(b []byte) (int, error) {
-	if frw.FileNotFound {
-		return len(b), nil
-	}
-	return frw.WrappedResponseWriter.Write(b)
-}
-
-// WriteHeader sends statusCode to wrapped response writer
-func (frw *FallbackResponseWriter) WriteHeader(statusCode int) {
-
-	if statusCode == http.StatusNotFound {
-		frw.FileNotFound = true
-		return
-	}
-
-	frw.WrappedResponseWriter.WriteHeader(statusCode)
-}
-
-func AddFallbackHandler(handler http.HandlerFunc, fs http.FileSystem) http.HandlerFunc {
+func addFallbackHandler(handler http.HandlerFunc, fs http.FileSystem) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		frw := FallbackResponseWriter{
 			WrappedResponseWriter: w,
